Rename 'users' test file to 'user' to match modules
