+++ code/api/src/__tests__/index_test.js
@@ -2,9 +2,9 @@
 import db from '../setup/database'
 //Tests to Run
 //Run Query Tests 1st
-import query from './modules/users/query_test'
+import query from './modules/user/query_test'
 //Run Mutation Tests 2nd
-import mutation from './modules/users/mutation_test'
+import mutation from './modules/user/mutation_test'
 
 afterAll(() => {
   db.close();
